[Breaking] `invites` plugin:
Initial commit of breaking changes.
- No longer store and retrieved user data on member join/leave.
- Files are now separated. Created new folder `core` and `models.py` in it.
